+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- [#785] Add explicit equality overloads for `RenderGroup.WithData<T>()`.
 
 ### Fixed
 
@@ -19,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 
 ### Deprecated
+- [#785] Deprecate `RenderGroup.WithData<T>(T)` because it uses heuristic equality rules.
 
 ## [1.12.0] - [2026-05-05]